wip: 컴파운드로 쪼개는 작업까지
