+++ apps/home/app/_shared/components/ActivityGraph/ActivityLogGraph.types.ts
@@ -1,0 +1,7 @@
+export type ActivityLogGraphProps = {
+  data: {
+    startDate: string
+    endDate: string
+    contents: string
+  }[]
+}
