+++ apps/home/app/_shared/components/ActivityGraph/compounds/ActivityLogGraph.WeekList.tsx
@@ -1,0 +1,23 @@
+import { cn } from '@hamsurang/ui'
+
+const getSrOnlyClass = ({ weekDay }: { weekDay: string }) => {
+  const isNonSrOnly = weekDay === 'Mon' || weekDay === 'Wed' || weekDay === 'Fri'
+  return isNonSrOnly ? '' : 'sr-only'
+}
+
+const weekDays = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
+
+export const ActivityLogGraphWeekList = () => {
+  return (
+    <aside className="flex flex-col justify-evenly">
+      {weekDays.map((weekDay) => (
+        <span
+          className={cn('text-xs text-right', getSrOnlyClass({ weekDay }))}
+          key={`${weekDay}+${crypto.randomUUID()}`}
+        >
+          {weekDay}
+        </span>
+      ))}
+    </aside>
+  )
+}
