+++ apps/home/app/_shared/components/ActivityGraph/compounds/index.ts
@@ -1,0 +1,4 @@
+export { ActivityLogGraphMonthList as MonthList } from './ActivityLogGraph.MonthList'
+export { ActivityLogGraphTable as Table } from './ActivityLogGraph.Table'
+export { ActivityLogGraphWeekList as WeekList } from './ActivityLogGraph.WeekList'
+export { ActivityLogGraphYearList as YearList } from './ActivityLogGraph.YearList'
